fix: Changed datatypes of pledge and cost field to fix sql error

+++ components/stores/staking/src/main/resources/db/migration/postgresql/V0_800_1__init.sql
@@ -68,8 +68,8 @@ create table pool_registration
     cert_index      int          not null,
     pool_id         varchar(255), -- pool hash
     vrf_key         varchar(255),
-    pledge          bigint,
-    cost            bigint,
+    pledge          numeric(20,0),
+    cost            numeric(20, 0),
     margin          decimal(10, 8),
     reward_account  varchar(255),
     pool_owners     jsonb,
